Add prelim scroll support using the J and K keys
The number 7 is being (before my participation) used as a hard-coded constant for
the max display-able entries. I have continued this, unfortunately for now.

Entries that do not equal or exceed the max number will not be scrollable (enforced)

an elipsis is printed above or below (or both) entries indicating there are more
in any given direction.