Make options script run its own copy of config
This removes the dependency on the background page (so we can tidy it
up) and also removes the slightly awkward message passing setup we had
there (where we had to be careful to postMessage arrays, but not
primitive types).

Furthermore, it's more inline with an event page or ServiceWorker
architecture so it's probably a step in the right direction for future
refactoring.