Drop RikaiBackground class
It doesn't add any value but just makes it harder to separate out the
different functionality.

If we ever decide to switch to an event page model (and that seems
unlikely--we should just jump straight to Service Workers), then we'll
need to move these event registrations to the top-level anyway.

More likely, if we try to separate out this logic into different modules
having it all wrapped in one class will just complicate things.

By splitting this out, it also makes it easier to do an initial callback
to db state listeners which this patch also does.