Differentiate version fields based on browser
Firefox allows alpha characters in the version number but does not
support the version_name field.

Chrome does not allow alpha characters but supports the version_name
field.

As a result we produce different versions based on the browser.

Once it comes time to release a new version, however, we should drop the
'pre' / 'alpha' designation. It's mostly there so we can ignore any
bugsnag reports that come in during development.