Add support for continuous forms
This fixes #73.

This also updates the types for masu stems so that we don't try to
further deinflect the result of a masu stem conjugation (since the
result should already be in the plain form).

This was necessary because otherwise we would deinflect 食べて into
食べてる (treating 食べて as a masu stem) and then proceed to de-inflect
that again using the continuous form into 食べる so the reason presented
for many て forms erroneously included the continuous form as a possible
conjugation reason.