Adds back Heisig 6th reference codes to kanji dictionary (#197)
Previously this information was manually added but it turns out it was
already in the kanjidic output. At that time, the contributor chose
LF to represent this code but since it's already defined as DN we will
change the implementation to match the dictionary file.

Updates kanji.dat accordingly.

* Change supported ref types to work with any length ref types.

Used a prefix match instead of a set look up since there's no way to
distinguish where the ref type ends.