fix: Do deep equality comparison for textAtPoint objects
Safari seems to generate a bunch of mousemove events when the Shift key
is released, some of which have shiftKey set to true, and others of
which don't (so we can't just ignore the ones with shiftKey = true).

That's not necessary a problem except that if our check to see if our
current text position has changed returns false positives we'll end up
doing a new search on the default dictionary meaning the result will
reset to the word dictionary each time we release the Shift key making
it basically impossible to cycle through dictionaries with the Shift
key.