chore: Fix warning about eval code
Webpack was injecting code into the options and background scripts to
detect the publicPath on the fly. This included a routine for
establishing the global object that included, in one branch, the code:

  new Function("return this");

This is turn was causing webext-lint to report a warning about eval
code.

By simply specifying the publicPath, however, we this code is no longer
generated and the warning is fixed.