chore: update to TS 4.6.2
In order to work around some lib.d.ts changes, I bumped the target and
lib versions in tsconfig.json.

This risks breaking users on older versions of Firefox if we start
generating code that is not supported there so I reviewed the generated
code and found the main differences seem to be:

- Use of spread operator in object literals (Firefox 55+)
- Use of unbound catch (Firefox 58+)

So hopefully this is fine.