fix: reflect source kana usage when displaying matches
Fixes #978.

Specifically when looking up words like バカ売れ and ガタガタ we would
prioritize the hiragana version, sometimes even hiding the katakana
version altogether.

This is because, when doing a search on the IDB database, we first
normalize the text to hiragana.

Now, the IDB database itself is able to look up kana variants but I
believe we were doing the hiragana conversion separately to that so we
can do deinflection using normalized hiragana.

This patch moves the normalization to the point where we do deinflection
so that we preserve the source kana form and jpdict-idb can use that to
more accurately set the match labels on the different headwords.

As a result, cases like サボった now correctly show サボる as the match
while cases that use katakana in the deinflected part should still work.