chore: revert Chrome and Edge to MV2 for now
At least in local development, the script we inject to Google Docs to
trigger annotated canvas seems to be blocked with MV3 due to changes to
CSP settings.

Maybe this doesn't happen for a properly published extension or maybe
I need to tweak the CSP settings, I'm not sure.

For now MV2 works so we'll stick with that.