feat: remove browser_styles from option, add dark mode
Firefox will no longer support browser_styles as of MV3 and it was
always problematic to begin with (it was hard or impossible to extend
and it means we have to test every change to the options page on both
Firefox and browsers that don't support browsers_styles separately).

While reworking the styles I've also made them support dark mode and
fixed a number of layout and useability issues.