chore: convert database summary to Preact/Tailwind (#1364)
We are getting more and more requests to add options and the current code which
updates the DOM for the options page using vanilla JS has reached its breaking
point.

I've tried various frameworks but finally settled on Preact. Svelte looked
promising but has a really hard time with types for event handlers. We'd like
not to regress type safety for this. Also, the only suitable component mounting
framework I could find for it was Storybook which wanted us to install too many
other dependencies.

I looked into Lit too but Web Components are still really tricky with regards to
accessibility and extracting styles.

Preact is very lightweight (more than SolidJS etc.) so we can probably use it in
the content script eventually without too much trouble. It's also a familiar
API for anyone wanting to contribute.

I probably didn't need to introduce Tailwind--the styles for the options page
are not too bad--but I misunderstood how React Cosmos was loading styles and
thought it was necessary to add some sort of CSS-in-JS solution. I tried to use
Vanilla Extract and various others approaches but none of them would let me both
colocate the styles with the component code and extract a single stylesheet at
the end.

All this is not only groundwork for making the options page easier to maintain
and extend but also we ultimately want to apply a similar approach to the popup.
This should allow us to drop all the half-baked test files we currently use and
finally make the popup a bit more interactive (e.g. clickable links etc.).

Furthermore, the popup styles are a mess so using something like Tailwind will
eventually allow us to maintain those styles without breaking things.

What's more, as part of doing this we had to introduce an approach for toggling
the locale. Once we use this approach everywhere we will finally be able to
allow the user to set the language used by 10ten's interface to something
different to their browser language.