fix: fix tab manager initialization race condition
STR:

1. Start the add-on but don't enable it
2. Wait for the download to finish
3. Terminate the background process from about:debugging (or just wait
   for it to terminate)
4. On any regular Web page, enable the add-on by clicking the icon on the
   toolbar

Expected results:

The add-on is enabled immediately.

Actual results:

It takes two clicks.

Analysis:

We'd hit a race condition where we'd call tabManager.init() while the
tabManager was still being initialized. As a result it would update its
this.enabled member to `true` but then the initialization steps would
reset it back to false.

The fix is just to wait for the initialization to complete before trying
to toggle the value.

As far as I can tell this does not affect the ActiveTabManager because
its initialization is synchronous.

Fixes #1491.