chore: add breadcrumbs for certain high-level operations
We've been getting an alarming number of AbortErrors being posted
recently but we have absolutely no context to work out where they are
coming from.

This patch adds some breadcrumbs to certain high-level operations in the
hope it helps to narrow down the cause of these errors.

Nothing is sent when there are no errors and input text is redacted when
it is sent.