fix: prevent starting a new reason chain with Reason.None
This should not be possible, as Reason.None cannot stand on its own and
needs a preceding rule to make sense.

This is a preventive measure, as this could lead to problems in the
future if one rule being a subset of another rule in input as well as
output is added, e.g. the rule
　たまう → 　たまう (te-form) being a subset of
のたまう → のたまう (te-form).