fix: pass in shadowRoots to caretPositionFromPoint
The spec for caretPositionFromPoint has been updated to require
providing `shadowRoots` in order to dig into Shadow DOM content.[1]

This has been implemented in Chrome[2] and is due to ship in 128 (in beta
at the time of this patch).

It has also been implemented in Firefox[3] but is currently only enabled on
the Nightly channel.[4]

This API is quite annoying for extensions that simply want to scan all
the text in the document but I guess it makes sense for regular Web
apps.

[1] https://github.com/w3c/csswg-drafts/issues/9932
[2] https://issues.chromium.org/issues/41117286
[3] https://bugzilla.mozilla.org/show_bug.cgi?id=1889503
[4] https://searchfox.org/mozilla-central/rev/64f196824745c5db5ef7bb23725cfb9a41586149/modules/libpref/init/StaticPrefList.yaml#4742-4751