Revert filtering popup from puck target elements
Previously, we were filtering the popup window element itself from the
possible target elements. This was to prevent the puck from holding the
popup open, which can happen when moving the puck very fast on iOS, and
interfere with scanning; however, this had the unfortunate side effect
that, when the puck was covering a part of the popup (which happens
semi-frequently depending on screen size), you could no longer move it
out of the way to see the covered area without the popup itself
disappearing.

This commit reverts that change.